refactor: - remove button fallback from ErrorPrompt component - create separate component for StartNewConversationButton
